backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
websockets==14.1
tomli==2.0.1
toml==0.10.2
any-llm-sdk[anthropic,openai]
httpx-aiohttp>=0.1.8
composio-core==0.7.20